Fold sendGalleryImagePost into messageSender
